+++ src/byte.hpp
@@ -15,8 +15,7 @@ class Byte {
 
         OpCode op;
 
-        int int_value;
-        std::string string_value;
+        std::string value;
 
     public:
 
@@ -27,11 +26,8 @@ class Byte {
         Byte(OpCode op)
         : op{ op } {}
 
-        Byte(OpCode op, int value)
-        : op{ op }, int_value{ value } {}
-
         Byte(OpCode op, std::string value)
-        : op{ op }, string_value{ value } {}
+        : op{ op }, value{ value } {}
 
         // Accessors
 
@@ -39,23 +35,19 @@ class Byte {
             return this->op;
         }
 
-        int get_int_value(void) {
-            return this->int_value;
-        }
-
-        std::string get_string_value(void) {
-            return this->string_value;
+        std::string get_value(void) {
+            return this->value;
         }
 
         // Helpers
 
         const std::string to_string(void) {
-            std::string out = "BYTE [OpCode: " + op_string.at(this->op);
+            std::string out = "BYTE [OpCode: " + op_code_string.at(this->op);
 
             if (this->op == OpCode::OP_IDENTIFIER) {
-                out += ", Value: " + this->string_value;
+                out += ", Value: " + this->value;
             } else if (this->op == OpCode::OP_CONSTANT) {
-                out += ", Value: " + std::to_string(this->int_value);
+                out += ", Value: " + this->value;
             }
 
             return out + "]";