+++ src/debug.hpp
@@ -4,7 +4,12 @@
 
 // #define DEBUG_SCANNER
 #define DEBUG_PRINT_TOKENS
+
 // #define DEBUG_PARSER
-#define DEBUG_PRINT_CHUNK
+#define DEBUG_PRINT_BYTES
+
+// #define DEBUG_TACKY
+#define DEBUG_PRINT_TACKY_BYTES
+
 // #define DEBUG_COMPILER
 // #define DEBUG_PRINT_INSTRUCTIONS