+++ src/tacky.hpp
@@ -1,0 +1,3 @@
+#include "byte.hpp"
+
+int generate_tacky(Byte *current_byte);